+++ pkg/macros/macros.go
@@ -6,7 +6,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/grafana/sqlds/v2"
+	"github.com/grafana/sqlds/v3"
+
+	"github.com/grafana/grafana-plugin-sdk-go/data/sqlutil"
 )
 
 // Converts a time.Time to a Date
@@ -25,26 +27,26 @@ func timeToDateTime64(t time.Time) string {
 }
 
 // FromTimeFilter returns a time filter expression based on grafana's timepicker's "from" time in seconds
-func FromTimeFilter(query *sqlds.Query, args []string) (string, error) {
+func FromTimeFilter(query *sqlutil.Query, args []string) (string, error) {
 	return timeToDateTime(query.TimeRange.From), nil
 }
 
 // ToTimeFilter returns a time filter expression based on grafana's timepicker's "to" time in seconds
-func ToTimeFilter(query *sqlds.Query, args []string) (string, error) {
+func ToTimeFilter(query *sqlutil.Query, args []string) (string, error) {
 	return timeToDateTime(query.TimeRange.To), nil
 }
 
 // FromTimeFilterMs returns a time filter expression based on grafana's timepicker's "from" time in milliseconds
-func FromTimeFilterMs(query *sqlds.Query, args []string) (string, error) {
+func FromTimeFilterMs(query *sqlutil.Query, args []string) (string, error) {
 	return timeToDateTime64(query.TimeRange.From), nil
 }
 
 // ToTimeFilterMs returns a time filter expression based on grafana's timepicker's "to" time in milliseconds
-func ToTimeFilterMs(query *sqlds.Query, args []string) (string, error) {
+func ToTimeFilterMs(query *sqlutil.Query, args []string) (string, error) {
 	return timeToDateTime64(query.TimeRange.To), nil
 }
 
-func TimeFilter(query *sqlds.Query, args []string) (string, error) {
+func TimeFilter(query *sqlutil.Query, args []string) (string, error) {
 	if len(args) != 1 {
 		return "", fmt.Errorf("%w: expected 1 argument, received %d", sqlds.ErrorBadArgumentCount, len(args))
 	}
@@ -58,7 +60,7 @@ func TimeFilter(query *sqlds.Query, args []string) (string, error) {
 	return fmt.Sprintf("%s >= %s AND %s <= %s", column, timeToDateTime(from), column, timeToDateTime(to)), nil
 }
 
-func TimeFilterMs(query *sqlds.Query, args []string) (string, error) {
+func TimeFilterMs(query *sqlutil.Query, args []string) (string, error) {
 	if len(args) != 1 {
 		return "", fmt.Errorf("%w: expected 1 argument, received %d", sqlds.ErrorBadArgumentCount, len(args))
 	}
@@ -72,7 +74,7 @@ func TimeFilterMs(query *sqlds.Query, args []string) (string, error) {
 	return fmt.Sprintf("%s >= %s AND %s <= %s", column, timeToDateTime64(from), column, timeToDateTime64(to)), nil
 }
 
-func DateFilter(query *sqlds.Query, args []string) (string, error) {
+func DateFilter(query *sqlutil.Query, args []string) (string, error) {
 	if len(args) != 1 {
 		return "", fmt.Errorf("%w: expected 1 argument, received %d", sqlds.ErrorBadArgumentCount, len(args))
 	}
@@ -85,7 +87,7 @@ func DateFilter(query *sqlds.Query, args []string) (string, error) {
 	return fmt.Sprintf("%s >= %s AND %s <= %s", column, timeToDate(from), column, timeToDate(to)), nil
 }
 
-func TimeInterval(query *sqlds.Query, args []string) (string, error) {
+func TimeInterval(query *sqlutil.Query, args []string) (string, error) {
 	if len(args) != 1 {
 		return "", fmt.Errorf("%w: expected 1 argument, received %d", sqlds.ErrorBadArgumentCount, len(args))
 	}
@@ -94,7 +96,7 @@ func TimeInterval(query *sqlds.Query, args []string) (string, error) {
 	return fmt.Sprintf("toStartOfInterval(toDateTime(%s), INTERVAL %d second)", args[0], int(seconds)), nil
 }
 
-func TimeIntervalMs(query *sqlds.Query, args []string) (string, error) {
+func TimeIntervalMs(query *sqlutil.Query, args []string) (string, error) {
 	if len(args) != 1 {
 		return "", fmt.Errorf("%w: expected 1 argument, received %d", sqlds.ErrorBadArgumentCount, len(args))
 	}
@@ -103,7 +105,7 @@ func TimeIntervalMs(query *sqlds.Query, args []string) (string, error) {
 	return fmt.Sprintf("toStartOfInterval(toDateTime64(%s, 3), INTERVAL %d millisecond)", args[0], int(milliseconds)), nil
 }
 
-func IntervalSeconds(query *sqlds.Query, args []string) (string, error) {
+func IntervalSeconds(query *sqlutil.Query, args []string) (string, error) {
 	seconds := math.Max(query.Interval.Seconds(), 1)
 	return fmt.Sprintf("%d", int(seconds)), nil
 }